backend/requirements.txt
Django==4.2.16
Djangorestframework==3.15.2
psycopg[binary]==3.1.8
django-cors-headers==4.4.0
django-allauth
channels==4.1.0
daphne==4.1.2
pillow==10.4.0
requests==2.26.0
dj-rest-auth
djangorestframework-simplejwt==5.0.0
pyotp==2.6.0
qrcode==7.2
channels-redis==4.2.0
python-dotenv===1.0.1
cloudinary==1.26.0
graypy==2.1.0
python-json-logger==2.0.7
